+++ src/MicroElements.Metadata.SystemTextJson/SystemTextJson/PropertyContainerConverterFactory.cs
@@ -1,4 +1,7 @@
-﻿using System;
+﻿// Copyright (c) MicroElements. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System;
 using System.Text.Json;
 using System.Text.Json.Serialization;
 using MicroElements.Functional;
@@ -7,7 +10,7 @@ using MicroElements.Metadata.Serialization;
 namespace MicroElements.Metadata.SystemTextJson
 {
     /// <summary>
-    /// Creates strong typed instances of <see cref="PropertyContainerConverter"/>.
+    /// Creates strong typed instances of <see cref="PropertyContainerConverter{TPropertyContainer}"/>.
     /// </summary>
     public class PropertyContainerConverterFactory : JsonConverterFactory
     {
@@ -34,7 +37,10 @@ namespace MicroElements.Metadata.SystemTextJson
         /// <inheritdoc />
         public override JsonConverter CreateConverter(Type typeToConvert, JsonSerializerOptions options)
         {
-            JsonConverter converter = new PropertyContainerConverter(Options, typeToConvert);
+            // JsonConverter converter = new PropertyContainerConverter<T>(Options);
+            Type genericConverterType = typeof(PropertyContainerConverter<>).MakeGenericType(typeToConvert);
+            JsonConverter converter = (JsonConverter)Activator.CreateInstance(genericConverterType, Options);
+
             return converter;
         }
     }